Test: refactor test suite into a dedicated folder
