+++ Test/unit_test.py
@@ -1,8 +1,6 @@
 import unittest
 from unittest import mock
 import io
-import sys
-import copy
 
 import responses
 
@@ -58,29 +56,9 @@ class TestFilterElement(unittest.TestCase):
 
 
 class TestGetCsvData(unittest.TestCase):
-    def setUp(self):
-        self.fake_websites_data = io.StringIO(
-            ",hash,filter,last_change_date\n"
-            "https://apple.com,"
-            "1ec1ea5fa8cd4e378f03dcce8b61667c19bef3343a4160ab122af04fbc7a9f6f,"
-            "page-home ac-nav-overlap,2020-10-01\n"
-            "https://www.provincia.brescia.it/istituzionale/concorsi,"
-            "714120cf1abda6bac178b4e23c9c2fbacc93b87c5b9b560d90af4e9904dc3be4,"
-            "col-sm-9,2020-05-21"
-        )
-        self.fake_websites_data_comment = io.StringIO(
-            ",hash,filter,last_change_date\n"
-            "#https://apple.com,"
-            "1ec1ea5fa8cd4e378f03dcce8b61667c19bef3343a4160ab122af04fbc7a9f6f,"
-            "page-home ac-nav-overlap,2020-10-01\n"
-            "https://www.provincia.brescia.it/istituzionale/concorsi,"
-            "714120cf1abda6bac178b4e23c9c2fbacc93b87c5b9b560d90af4e9904dc3be4,"
-            "col-sm-9,2020-05-21"
-        )
-
     def test_get_csv_data(self):
         self.assertEqual(
-            get_csv_data(self.fake_websites_data),
+            get_csv_data(pathlib.Path(__file__).with_name("website_data.csv")),
             {
                 'https://apple.com': {
                     'hash': '1ec1ea5fa8cd4e378f03dcce8b61667c19bef334'
@@ -99,7 +77,9 @@ class TestGetCsvData(unittest.TestCase):
 
     def test_get_csv_data_with_comments(self):
         self.assertEqual(
-            get_csv_data(self.fake_websites_data_comment),
+            get_csv_data(
+                pathlib.Path(__file__).with_name("website_data_comment.csv")
+            ),
             {
                 'https://www.provincia.brescia.it/istituzionale/concorsi': {
                     'hash': '714120cf1abda6bac178b4e23c9c2fbacc93b87c5'
@@ -185,64 +165,6 @@ class TestOpenWebsite(unittest.TestCase):
         response_content = open_website('http://www.test.com')
 
         self.assertEqual(response_content, b'Hello world!')
-
-
-class TestPerformCheck(unittest.TestCase):
-    # actually this is an Integration test
-    def setUp(self):
-        # create websites to monitor,
-        # the first will change, the second won't
-        self.fake_website_data = {
-            'http://www.changed.com': {
-                'hash': 'fakehash',
-                'filter': 'my_id',
-                'last_change_date': '2020-01-01',
-            },
-            'http://www.unchanged.com': {
-                'hash': '10501de66339424ad03fe30cdc4d8ea6246f81aa65065493'
-                        'fb2692599717aa6f',
-                'filter': 'my_class',
-                'last_change_date': datetime.date.today().isoformat()
-            }
-        }
-
-        # set the responses of the fake websites
-        responses.add(
-            method=responses.GET, url='http://www.changed.com',
-            body=b"<!DOCTYPE html><html><body>"
-                 b"<p id='my_id'>This is a paragraph.</p>"
-                 b"</body></html>",
-            status=200
-        )
-        responses.add(
-            method=responses.GET, url='http://www.unchanged.com',
-            body=b"<!DOCTYPE html><html><body>"
-                 b"<p class='my_class'>hello world</p>"
-                 b"</body></html>",
-            status=200
-        )
-
-    @responses.activate
-    def test_output(self):
-        changed = perform_check(copy.deepcopy(self.fake_website_data))
-
-        self.assertEqual(changed, ["http://www.changed.com changed!"])
-
-    @responses.activate
-    def test_changed_dict(self):
-        # deepcopy the fake dict to prevent it from being modified
-        copy_fake_website_data = copy.deepcopy(self.fake_website_data)
-        perform_check(copy_fake_website_data)
-
-        self.assertEqual(
-            copy_fake_website_data['http://www.changed.com'],
-            {
-                'hash': 'f7605b3e0e68c7823e4de3c2911f3c1a899ca0758d2d2222'
-                        'afd193e824ef0133',
-                'filter': 'my_id',
-                'last_change_date': datetime.date.today().isoformat()
-            }
-        )
 
 
 if __name__ == '__main__':
